+++ core-bundle/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### 4.4.0-beta1 (2017-05-05)
 
+ * Always add custom meta fields to the templates (see #717).
  * Warn if another user has edited a record when saving it (see #809).
  * Optimize the element preview height (see #810).
  * Use the file meta data by default when adding an image (see #807).