+++ core-bundle/CHANGELOG.md
@@ -1,9 +1,12 @@
 # Contao core bundle change log
 
-### 4.4.0-beta1 (2017-05-05)
+### DEV
 
  * Update the Punycode library to version 2 (see #748).
  * Always add custom meta fields to the templates (see #717).
+
+### 4.4.0-beta1 (2017-05-05)
+
  * Warn if another user has edited a record when saving it (see #809).
  * Optimize the element preview height (see #810).
  * Use the file meta data by default when adding an image (see #807).